Added functionality to let user decide where the scan report should be saved.

+++ krackPlusCLI.py
@@ -59,6 +59,8 @@ def main():
                       " Password length has to be 8 characters or more!", dest='password')
     parser.add_option('-d', help="This option will increase output verbosity for KRACK+ Scan", dest='debug', action='store_true')
     parser.add_option('--dd', help="This option will increase output verbosity even more for KRACK+ Scan (debugging purposes).", dest='dd', action='store_true')
+    parser.add_option('--path', '-p', help="Set path where scan report should be saved", dest='path')
+
 
     # General KRACK+ options:
     parser.add_option('--restore', '-r', help="This option will restore internet connection (wifi). Hopefully you'll never have to use this option.", dest='restore', default=False, action='store_true')
@@ -97,12 +99,17 @@ def main():
                 log.info("Cleaning up and generating PDF report of findings...")
                 subprocess.call(["./restoreClientWifi.sh"])
                 writeResults()
-                subprocess.call("./genPDF.py")
+		if options.path:
+                	subprocess.call(["./genPDF.py " + options.path], shell=True)
+			log.info("PDF generated in '" + options.path + "'.")
+		else: 
+			subprocess.call("./genPDF.py")	
+			log.info("PDF generated in '" + path + "'.")
                 subprocess.call(["rm scanOutput.txt"], shell=True)
                 subprocess.call(["rm scannedMacIP.txt"], shell=True)
                 subprocess.call(["rm pairwiseVulnMacIP.txt"], shell=True)
                 subprocess.call(["rm groupVulnMacIP.txt"], shell=True)
-                log.info("PDF generated in '" + path + "'.")
+                
         except:
             log.error("Error occurred.")
             log.info("Restoring internet connection.")
@@ -116,18 +123,21 @@ def main():
             #Sets up dependencies before the attack script runs
             subprocess.call(["./prepareClientAttack.sh"])
             with open('./attackOutput.txt', 'w') as attackOutput:
-                # Usually not necesary to cd to run a script, however Vanhoef's implementation requires it, otherwise we would have to alter his code. 
+                # Subprocess runs script from Vanhoef's repository, to avoid problems with the temporary files his script creates
                 subprocess.call(["cd krackattacks-poc-zerokey/krackattack/ && ./krack-all-zero-tk.py " + options.rogue + " " +
                                  options.mon + " " + options.targetSSID + " --target " + options.target + " &"], stdout=attackOutput, shell=True)
                 # Usually not necesary to cd to run a script, however Vanhoef's implementation requires it, otherwise we would have to alter his code. 
                 #subprocess.call(["cd krackattacks-poc-zerokey/krackattack/ && ./enable_internet_forwarding.sh &"])
                 #subprocess.call(["sslstrip -w sslstrip.log &"])
             	attackParser()
+
         except KeyboardInterrupt:
+   	    subprocess.call(["clear"], shell=True)
             log.info("Cleaning up and restoring wifi ...")
 	    subprocess.call(["rm attackOutput.txt"], shell=True)
             subprocess.call(["./restoreClientWifi.sh"])
 	except:
+	    subprocess.call(["clear"], shell=True)
             log.info("Error occurred. Restoring wifi ...")
             subprocess.call(["rm attackOutput.txt"], shell=True)
             subprocess.call(["./restoreClientWifi.sh"])
